postwall/requirements.txt
# Spisok zavisimostej

Flask
Flask-SQLAlchemy
Flask-Migrate
mysqlclient